Refactor converters classes
