+++ conversion/converters/mscoco.py
@@ -1,0 +1,14 @@
+def kek2mscoco():
+    pass
+
+
+def mscoco2kek():
+    pass
+
+
+def mscoco_simple2kek():
+    pass
+
+
+def mscoco_hard2kik():
+    pass
